Create a documentation directory
I think it is more common to use a `docs/` folder to organize the
documentation. It makes it easier to share/fork than when using the
built-in Wiki. It also allows to prepare the documentation in separate
branches (like for instance the new way to run the integration tests but
only starting from v11.0.0).

I also created a README.md that will show in the main page of the
project. It presents the project and links to the documentation pages.