Change: do not use the access token secret for protected resources
The standard is clear: only the access token is needed in order to get
protected resources, but RestSharp still asks for it.

I think it's a bug on their side (I tested OAuth without the token
secret for a couple of months now and had no problem).

As per their guidelines, I opened a thread on their mailing list:
https://groups.google.com/forum/#!topic/restsharp/DelqSWi6VPU

In the meantime, passing `string.Empty` works just as well.