Removes exception when library fails to discover the custom field type.
Fixes issue #481.

This addresses the problem when Jira changes its set of custom fields after the CustomFieldTypes have already been cached by the library.

With this change, when the library fails to locate the custom field type from the cache, it will default to using a string or string[] serializer.