Merged in upstream-code-cleaning-part-2 (pull request #100)
Convert to LF line endings

* Merged farmas/atlassian.net-sdk into master

* Merged farmas/atlassian.net-sdk into master

* Merged farmas/atlassian.net-sdk into master
* Update .editorconfig

All files:
* LF end of lines
* Insert final new line
* Trim trailing whitespace

Some exception for .md, .yml and .csproj

* Convert to LF line endings

**Why LF and not CRLF?**

Out of 193 C# files, 38 are CRLF or a mix of LF and CRLF. This
represents 20% of the code base. So the 80% of LF files wins.

The other files (.md, .yml, .txt, etc.) are using the same conventions
in order to simplify the code base guidelines.

Only exceptions:
* *.csproj and *.yml: only 2 spaces for indent size
* *.md: allow trailing whitespaces (it is part of Markdown standard)


Approved-by: Federico Silva Armas